refactor(display-list): split into separate modules (#184)
